style(cloud.project.billing): rename services file
add legacy suffix for legacy service
rename agora service to current service
